pyproject.toml
[tool.poetry]
name = "currency-test"
version = "0.1.0"
description = ""
authors = ["Slava Kuryaev <92830706+necrosskull@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.0"}
sqlalchemy = "^2.0.32"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"
asyncpg = "^0.29.0"
coloredlogs = "^15.0.1"
termcolor = "^2.4.0"
ruff = "^0.5.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
